# packages/codegen/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Multi-Tenant Shopify API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/orders/": {
      "get": {
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_orders__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Status"
            }
          },
          {
            "name": "payment_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Payment Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/OrderResponse" },
                  "title": "Response List Orders Api V1 Orders  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Order",
        "operationId": "create_order_api_v1_orders__post",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Order",
        "operationId": "update_order_api_v1_orders__order_id__put",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Order",
        "operationId": "patch_order_api_v1_orders__order_id__patch",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/purchase-orders": {
      "get": {
        "summary": "Get Order Purchase Orders",
        "description": "Return all purchase orders linked to a given sales order.",
        "operationId": "get_order_purchase_orders_api_v1_orders__order_id__purchase_orders_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssociatedPOResponse"
                  },
                  "title": "Response Get Order Purchase Orders Api V1 Orders  Order Id  Purchase Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssociatedPOResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "po_number": { "type": "string", "title": "Po Number" },
          "status": { "type": "string", "title": "Status" },
          "supplier_name": { "type": "string", "title": "Supplier Name" },
          "total": { "type": "integer", "title": "Total" },
          "fulfillment_strategy": {
            "type": "string",
            "title": "Fulfillment Strategy"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "po_number",
          "status",
          "supplier_name",
          "total",
          "fulfillment_strategy",
          "created_at"
        ],
        "title": "AssociatedPOResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreate": {
        "properties": {
          "customer_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Customer Id"
          },
          "order_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Order Number"
          },
          "status": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Status"
          },
          "items": {
            "items": { "$ref": "#/components/schemas/OrderItemCreate" },
            "type": "array",
            "title": "Items"
          },
          "subtotal": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Subtotal",
            "default": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "default": 0
          },
          "shipping_address": {
            "additionalProperties": true,
            "type": "object",
            "title": "Shipping Address"
          },
          "billing_address": {
            "additionalProperties": true,
            "type": "object",
            "title": "Billing Address"
          },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "OrderCreate"
      },
      "OrderItemCreate": {
        "properties": {
          "variant_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Variant Id"
          },
          "product_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name",
            "default": ""
          },
          "sku": { "type": "string", "title": "Sku", "default": "" },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          },
          "unit_price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unit Price"
          },
          "total_price": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Price",
            "default": 0
          },
          "discount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Discount",
            "default": 0
          }
        },
        "type": "object",
        "required": ["quantity", "unit_price"],
        "title": "OrderItemCreate"
      },
      "OrderItemResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "variant_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Variant Id"
          },
          "product_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Product Id"
          },
          "product_name": { "type": "string", "title": "Product Name" },
          "variant_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Variant Name"
          },
          "sku": { "type": "string", "title": "Sku" },
          "quantity": { "type": "integer", "title": "Quantity" },
          "unit_price": { "type": "integer", "title": "Unit Price" },
          "total_price": { "type": "integer", "title": "Total Price" },
          "discount": { "type": "integer", "title": "Discount" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "product_name",
          "sku",
          "quantity",
          "unit_price",
          "total_price",
          "discount",
          "created_at"
        ],
        "title": "OrderItemResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "customer_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Customer Id"
          },
          "order_number": { "type": "string", "title": "Order Number" },
          "status": { "type": "string", "title": "Status" },
          "payment_status": { "type": "string", "title": "Payment Status" },
          "payment_method": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Payment Method"
          },
          "payment_intent_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Payment Intent Id"
          },
          "subtotal": { "type": "integer", "title": "Subtotal" },
          "tax": { "type": "integer", "title": "Tax" },
          "shipping": { "type": "integer", "title": "Shipping" },
          "discount": { "type": "integer", "title": "Discount" },
          "total": { "type": "integer", "title": "Total" },
          "currency": { "type": "string", "title": "Currency" },
          "shipping_address": {
            "additionalProperties": true,
            "type": "object",
            "title": "Shipping Address"
          },
          "billing_address": {
            "additionalProperties": true,
            "type": "object",
            "title": "Billing Address"
          },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          },
          "customer_email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Email"
          },
          "options": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Options"
          },
          "items": {
            "items": { "$ref": "#/components/schemas/OrderItemResponse" },
            "type": "array",
            "title": "Items"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "order_number",
          "status",
          "payment_status",
          "subtotal",
          "tax",
          "shipping",
          "discount",
          "total",
          "currency",
          "shipping_address",
          "billing_address",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "OrderResponse"
      },
      "OrderUpdate": {
        "properties": {
          "status": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Status"
          },
          "payment_status": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Payment Status"
          },
          "shipping_address": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Shipping Address"
          },
          "billing_address": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Billing Address"
          },
          "notes": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "OrderUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/admin/auth/login" }
        }
      }
    }
  }
}